BR-Helpdesk/requirements2.txt
Flask>=0.12.2
matplotlib>=0.0
sklearn>=0.0
numpy>=0.0
pandas_ml>=0.0
nltk>=0.0
gensim>=2.1
gunicorn>=19.7.1
PyMySQL>=0.8.0
Flask-PyMongo>=0.5.1
Flask-SQLAlchemy>=2.3.2
Flask-PyMongo>=0.5.1
PyMongo>=3.6.0
six>=1.11.0
google-cloud-datastore>=1.6.0
google-cloud-storage>=0.0
textacy>=0.0
dash>=0.21.1 
dash-renderer>=0.13.0 
dash-html-components>=0.11.0 
dash-core-components>=0.24.0  
plotly>=0.0